backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.2
pydantic-settings>=2.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-multipart>=0.0.6
celery>=5.3.6
redis>=5.0.1
pdfplumber>=0.10.3
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
